# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MetaSRS Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Classes ─────────────────────────────────────────────────────────────────
...
  /reviews:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Submit a card review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /students/{studentId}/due-cards:
    get:
      operationId: getDueCards
      tags: [reviews]
      summary: Get cards due for review for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: deckId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Due cards with SRS state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DueCard"
  /students/{studentId}/reviews:
    get:
      operationId: listStudentReviews
      tags: [reviews]
      summary: List review history for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: deckId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Review history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ─── Analytics ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Users ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [teacher, student]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [teacher, student]
      required:
        - name
        - email
        - role

    # ─── Classes ───────────────────────────────────────────────────────────────
...
    Review:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        cardId:
          type: integer
        deckId:
          type: integer
        grade:
          type: integer
          description: "1=Again, 2=Hard, 3=Good, 4=Easy"
        recalled:
          type: boolean
        elapsedDays:
          type: number
        stabilityBefore:
          type: ["number", "null"]
        stabilityAfter:
          type: ["number", "null"]
        difficultyAfter:
          type: ["number", "null"]
        nextReviewAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - cardId
        - deckId
        - grade
        - recalled
        - elapsedDays
        - nextReviewAt
        - reviewedAt
    CreateReviewBody:
      type: object
      properties:
        studentId:
          type: integer
        cardId:
          type: integer
        deckId:
          type: integer
        grade:
          type: integer
          description: "1=Again, 2=Hard, 3=Good, 4=Easy"
        elapsedDays:
          type: number
      required:
        - studentId
        - cardId
        - deckId
        - grade
        - elapsedDays
    DueCard:
      type: object
      properties:
        cardId:
          type: integer
        deckId:
          type: integer
        front:
          type: string
        back:
          type: string
        hint:
          type: ["string", "null"]
        stabilityDays:
          type: ["number", "null"]
        difficulty:
          type: ["number", "null"]
        predictedRetention:
          type: ["number", "null"]
        reviewCount:
          type: integer
        lastReviewedAt:
          type: ["string", "null"]
          format: date-time
        nextReviewAt:
          type: ["string", "null"]
          format: date-time
        isNew:
          type: boolean
      required:
        - cardId
        - deckId
        - front
        - back
        - reviewCount
        - isNew

    # ─── Analytics ─────────────────────────────────────────────────────────────
...