# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MetaSRS Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  studentId: number;
  cardId: number;
  deckId: number;
  /** 1=Again, 2=Hard, 3=Good, 4=Easy */
  grade: number;
  recalled: boolean;
  elapsedDays: number;
  /** @nullable */
  stabilityBefore?: number | null;
  /** @nullable */
  stabilityAfter?: number | null;
  /** @nullable */
  difficultyAfter?: number | null;
  nextReviewAt: Date;
  reviewedAt: Date;
}